requirements.txt
flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
Pillow==10.2.0
requests==2.31.0
flask-cors==4.0.0
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
urllib3==2.2.0
opencv-python==4.9.0.80
numpy==1.26.4
redis==5.0.1
reportlab==4.1.0
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
Flask-Limiter[redis]==3.5.0 